tokens_generated_chart/requirements.txt
requests
pandas
matplotlib
python-dotenv
seaborn
dune-client